Added printer and nfc reader status on admin page
To get the printer status, the current implementation is a bit of a hack just to get something working.
Right now it uses the lpstat and lpoptions commands and parses the returned text.
The longer-term approach would be to find or write a C/C++ adapter that interfaces with
the CUPS API to get these values directly.